% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingscores.R
\docType{data}
\name{wingscores}
\alias{wingscores}
\title{Wing scores for ~7500 individual flies from 2018 neural defects course}
\format{An object of class \code{data.frame} with 7754 rows and 16 columns.}
\usage{
wingscores
}
\description{
Wing scores for ~7500 individual flies from 2018 neural defects course
}
\keyword{datasets}
