% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingscores_s1.R
\docType{data}
\name{wingscores_s1}
\alias{wingscores_s1}
\title{Wing scores for student 1}
\format{An object of class \code{data.frame} with 60 rows and 7 columns.}
\usage{
wingscores_s1
}
\description{
Wing scores for student 1
}
\keyword{datasets}
