% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingscores_s1F.R
\docType{data}
\name{wingscores_s1F}
\alias{wingscores_s1F}
\title{Wing scores for 24 females flies from student 1}
\format{An object of class \code{data.frame} with 24 rows and 7 columns.}
\usage{
wingscores_s1F
}
\description{
Wing scores for 24 females flies from student 1
}
\keyword{datasets}
