% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wngscrs_sxn1.R
\docType{data}
\name{wngscrs_sxn1}
\alias{wngscrs_sxn1}
\title{Wing scores for section 1}
\format{An object of class \code{data.frame} with 624 rows and 15 columns.}
\usage{
wngscrs_sxn1
}
\description{
Wing scores for section 1
}
\keyword{datasets}
