% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wngscrs_sxn1_able.R
\docType{data}
\name{wngscrs_sxn1_able}
\alias{wngscrs_sxn1_able}
\title{Wing scores for 3 students 3 alleles of the able gene}
\format{An object of class \code{data.frame} with 159 rows and 15 columns.}
\usage{
wngscrs_sxn1_able
}
\description{
Wing scores for 3 students 3 alleles of the able gene
}
\keyword{datasets}
