% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shroom-package.R
\docType{package}
\name{shroom-package}
\alias{shroom}
\alias{shroom-package}
\title{shroom: Example Analyses for Cell and Molecular Biology}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
